@@ -4,20 +4,25 @@
   "scripts": {
     "test": "jest",
     "test-watch": "jest --watch",
-    "lint": "prettier app/packs --check && eslint app/packs && stylelint 'app/packs/**/*.scss'",
-    "lint-fix": "prettier app/packs --write; eslint --fix app/packs; stylelint --fix 'app/packs/**/*.scss'"
+    "lint": "prettier app/javascript --check && eslint app/javascript && stylelint 'app/javascript/**/*.scss'",
+    "lint-fix": "prettier app/javascript --write; eslint --fix app/javascript; stylelint --fix 'app/javascript/**/*.scss'"
   },
   "dependencies": {
+    "@babel/core": "^7.16.10",
+    "@babel/plugin-transform-runtime": "^7.16.10",
+    "@babel/preset-env": "^7.16.10",
     "@babel/preset-react": "^7.16.0",
     "@babel/preset-typescript": "^7.16.7",
+    "@babel/runtime": "^7.17.9",
     "@heroicons/react": "^1.0.6",
     "@rails/actioncable": "^7.0.0",
     "@rails/activestorage": "^7.0.0",
     "@rails/ujs": "^7.0.0",
-    "@rails/webpacker": "^6.0.0-rc.6",
-    "@testing-library/user-event": "^13.5.0",
+    "autoprefixer": "^10.4.6",
+    "babel-loader": "^8.2.5",
     "classnames": "^2.3.1",
     "color-string": "^1.9.0",
+    "compression-webpack-plugin": "^9.2.0",
     "css-loader": "^6.5.1",
     "css-minimizer-webpack-plugin": "^3.2.0",
     "dot-prop": "^7.2.0",
@@ -26,49 +31,43 @@
     "postcss": "^8.3.9",
     "postcss-import": "^14.0.2",
     "postcss-loader": "^6.2.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "sass": "^1.45.0",
     "sass-loader": "^12.1.0",
+    "shakapacker": "^6.0.0",
     "style-loader": "^3.3.1",
     "tailwindcss": "^2.2.17",
+    "terser-webpack-plugin": "^5.3.1",
     "typescript": "^4.5.4",
-    "webpack": "^5.51.1",
-    "webpack-cli": "^4.9.2"
+    "webpack": "^5.66.0",
+    "webpack-assets-manifest": "^5.1.0",
+    "webpack-cli": "^4.9.1",
+    "webpack-merge": "^5.8.0",
+    "webpack-sources": "^3.2.3"
   },
   "engines": {
     "node": "16.x",
     "yarn": "1.x"
   },
   "version": "0.1.1",
-  "babel": {
-    "plugins": [
-      [
-        "@babel/plugin-proposal-private-property-in-object",
-        {
-          "loose": true
-        }
-      ]
-    ],
-    "presets": [
-      "./node_modules/@rails/webpacker/package/babel/preset.js"
-    ]
-  },
   "browserslist": [
     "defaults"
   ],
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/dom": "^8.13.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^14.1.1",
     "@types/jest": "^27.0.2",
     "@types/rails__activestorage": "^7.0.0",
     "@types/rails__ujs": "^6.0.1",
-    "@types/react": "^17.0.43",
-    "@types/react-dom": "^17.0.14",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
     "@types/testing-library__jest-dom": "^5.14.3",
     "@types/webpack-env": "^1.16.2",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.31.2",
+    "@typescript-eslint/eslint-plugin": "^5.21.0",
+    "@typescript-eslint/parser": "^5.21.0",
     "babel-jest": "^27.4.5",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
